+++ tests/Paramore.Brighter.Core.Tests/CommandProcessors/When_An_Exception_Is_Thrown_Terminate_The_Pipeline.cs
@@ -27,6 +27,7 @@ using FluentAssertions;
 using Paramore.Brighter.Core.Tests.CommandProcessors.TestDoubles;
 using Polly.Registry;
 using Microsoft.Extensions.DependencyInjection;
+using Paramore.Brighter.Core.Tests.TestHelpers;
 using Paramore.Brighter.Extensions.DependencyInjection;
 using Xunit;