+++ tests/Paramore.Brighter.Core.Tests/ControlBus/When_we_build_a_control_bus_we_can_send_configuration_messages_to_it.cs
@@ -25,6 +25,7 @@ THE SOFTWARE. */
 using System;
 using FakeItEasy;
 using FluentAssertions;
+using Paramore.Brighter.Core.Tests.TestHelpers;
 using Xunit;
 using Paramore.Brighter.ServiceActivator;
 using Paramore.Brighter.ServiceActivator.ControlBus;