+++ tests/Paramore.Brighter.Core.Tests/ExceptionPolicy/When_A_Fallback_Is_Broken_Ciruit_Only.cs
@@ -30,6 +30,7 @@ using Xunit;
 using Paramore.Brighter.Policies.Handlers;
 using Polly.Registry;
 using Microsoft.Extensions.DependencyInjection;
+using Paramore.Brighter.Core.Tests.TestHelpers;
 using Paramore.Brighter.Extensions.DependencyInjection;
 
 namespace Paramore.Brighter.Core.Tests.ExceptionPolicy