+++ tests/Paramore.Brighter.Core.Tests/MessageSerialisation/When_Wrapping_A_Message_Mapper_But_Not_In_Transform_Factory.cs
@@ -1,5 +1,6 @@
 ﻿using FluentAssertions;
 using Paramore.Brighter.Core.Tests.MessageSerialisation.Test_Doubles;
+using Paramore.Brighter.Core.Tests.TestHelpers;
 using Xunit;
 
 namespace Paramore.Brighter.Core.Tests.MessageSerialisation;