+++ release_notes.md
@@ -7,6 +7,23 @@ This section lists features in master, available by [AppVeyor](https://ci.appvey
 
 ## Master ##
 
+## Binary Serialization Fixes
+
+* MessageBody  nows store the character encoding type (defaults to UTF8) to allow correct conversion back to a string when using Value property
+* Use a CharacterEncoding.Raw for binary content (will be a Base64 string for Value)
+* Kafka transport payload is now byte[] and not string. This prevents corruption of Kafka 'header' of 5 bytes to store schema registry when used with schema registry support
+* DynamoDb now uses a byte[] and not a string for the message body to prevent lossy conversions
+* ContentType on Header is set from Body, if not set on the Header
+
+## Kafka Fixes
+
+* Kafka now serliases the ReplyTo Header correctly
+
+## New Transforms
+
+* Compression Transform now available to compress messages using Gzip (or Brotli or Deflate on .NET 6 or 7)
+
+
 ## Release 9.3.6 ##
 
 - Set correct partition key (kafka key) for Kafka messages